ingest_data/requirements.txt
gspread
google-auth-oauthlib
boto3
google-auth-httplib2
pandas
google-api-python-client
python-dotenv
openpyxl
streamlit